+++ torch/__init__.py
@@ -108,40 +108,7 @@ def set_default_tensor_type(t):
     _C._set_default_tensor_type(Tensor)
 
 
-def set_rng_state(new_state):
-    r"""Sets the random number generator state.
-
-    Args:
-        new_state (torch.ByteTensor): The desired state
-    """
-    default_generator.set_state(new_state)
-
-
-def get_rng_state():
-    r"""Returns the random number generator state as a ByteTensor."""
-    return default_generator.get_state()
-
-
-def manual_seed(seed):
-    r"""Sets the seed for generating random numbers. And returns a
-    `torch._C.Generator` object.
-
-    Args:
-        seed (int or long): The desired seed.
-    """
-    if not torch.cuda._in_bad_fork:
-        torch.cuda.manual_seed_all(seed)
-
-    return default_generator.manual_seed(seed)
-
-
-def initial_seed():
-    r"""Returns the initial seed for generating random numbers as a
-    python `long`.
-    """
-    return default_generator.initial_seed()
-
-
+from .random import set_rng_state, get_rng_state, manual_seed, initial_seed
 from .serialization import save, load
 from ._tensor_str import set_printoptions
 
@@ -341,6 +308,7 @@ import torch.multiprocessing
 import torch.sparse
 import torch.utils.backcompat
 import torch.onnx
+import torch.random
 
 _C._init_names(list(torch._tensor_classes) + list(torch._storage_classes))
